refactor(app): isolated hooks and components in `App.vue`
A refactorization has been done in `App.vue`. The reason of doing so is because that the difficulty
to maintain this file has become huge since the code lines exceeded 680 lines, with nested
algorithms in it handling 6 features. This has already lead to huge maintainability issues this
afternoon when I was busy modifying the site for April Fools Day easter egg. So I have isolated this
into different mixins to allow the main entry of the app to use that and utilize the hooks within
multiple mixins.
