+++ .env
@@ -1,5 +1,3 @@
-VUE_APP_I18N_LOCALE=zh
-VUE_APP_I18N_FALLBACK_LOCALE=zh
 SENTRY_URL=https://sentry.io
 SENTRY_ORG=penguin-statistics
 SENTRY_PROJECT=frontend
