+++ src/components/toolbar/NetworkStateIndicator.vue
@@ -1,5 +1,54 @@
 <template>
   <div>
+    <v-slide-x-reverse-transition>
+      <v-card
+        v-if="show"
+        class="network-state-indicator transition-all"
+        :class="{'error': haveError, 'blue darken-2': !haveError}"
+      >
+        <v-fade-transition>
+          <span
+            v-if="haveError && !model"
+            style="cursor: pointer"
+            @click="model = true"
+          >
+            <v-progress-circular
+              v-if="pending"
+              indeterminate
+              color="white"
+              class="mr-1"
+              :size="16"
+              :width="2"
+            />
+            <v-icon
+              v-else
+              small
+              class="mr-1"
+            >
+              mdi-alert
+            </v-icon>
+
+            <span class="caption white--text">
+              {{ $t('fetch.failed.title') }} ({{ errors.length }})
+            </span>
+          </span>
+          <span
+            v-else-if="pending"
+          >
+            <v-progress-circular
+              indeterminate
+              color="white"
+              class="mr-1"
+              :size="16"
+              :width="2"
+            />
+            <span class="caption white--text">
+              {{ $t('meta.loading') }}
+            </span>
+          </span>
+        </v-fade-transition>
+      </v-card>
+    </v-slide-x-reverse-transition>
     <v-dialog
       v-model="model"
       width="600"
@@ -66,55 +115,6 @@
         </v-card-actions>
       </v-card>
     </v-dialog>
-    <v-slide-x-reverse-transition>
-      <v-card
-        v-if="show"
-        class="network-state-indicator transition-all"
-        :class="{'error': haveError, 'blue darken-2': !haveError}"
-      >
-        <v-fade-transition>
-          <span
-            v-if="haveError && !model"
-            style="cursor: pointer"
-            @click="model = true"
-          >
-            <v-progress-circular
-              v-if="pending"
-              indeterminate
-              color="white"
-              class="mr-1"
-              :size="16"
-              :width="2"
-            />
-            <v-icon
-              v-else
-              small
-              class="mr-1"
-            >
-              mdi-alert
-            </v-icon>
-
-            <span class="caption">
-              {{ $t('fetch.failed.title') }} ({{ errors.length }})
-            </span>
-          </span>
-          <span
-            v-else-if="pending"
-          >
-            <v-progress-circular
-              indeterminate
-              color="white"
-              class="mr-1"
-              :size="16"
-              :width="2"
-            />
-            <span class="caption">
-              {{ $t('meta.loading') }}
-            </span>
-          </span>
-        </v-fade-transition>
-      </v-card>
-    </v-slide-x-reverse-transition>
   </div>
 </template>
