+++ src/mixins/hooks/CrispCustomizer.js
@@ -1,0 +1,40 @@
+import Console from "@/utils/Console";
+
+export default {
+  watch: {
+    '$route': [
+      'crispOpacityChanger'
+    ],
+  },
+  created () {
+    window.$crisp.push(["on", "session:loaded", () => {
+      // resolve safe-area
+      Console.info("[crisp] triggered | chat:loaded")
+      try {
+        document.querySelector("div.crisp-client > div#crisp-chatbox > div > a").style.setProperty("bottom", "calc(max(env(safe-area-inset-bottom), 14px))", "important");
+        document.querySelector("div.crisp-client > div#crisp-chatbox > div > a > span:nth-child(2)").style.setProperty("box-shadow", "0 0 5px rgba(0, 0, 0, .4)", "important");
+        this.crispOpacityChanger()
+      } catch (e) {
+        Console.error("[crisp] failed to initialize custom style:", e)
+      }
+    }]);
+  },
+  methods: {
+    crispOpacityChanger (newRoute = this.$route) {
+      Console.info("[crisp] customize | changing opacity");
+      try {
+        document.querySelector("div.crisp-client").style.setProperty("transition", "all 275ms cubic-bezier(0.165, 0.84, 0.44, 1)", "important");
+
+        if (newRoute.name === "home") {
+          document.querySelector("div.crisp-client").style.setProperty("opacity", 1, "important");
+          // document.querySelector("div.crisp-client").style.setProperty("transform", "translateY(0px)", "important")
+        } else {
+          document.querySelector("div.crisp-client").style.setProperty("opacity", 0, "important");
+          // document.querySelector("div.crisp-client").style.setProperty("transform", "translateY(32px)", "important")
+        }
+      } catch (e) {
+        Console.error("failed to change crisp opacity", e, "newRoute.name", newRoute.name)
+      }
+    },
+  },
+}
