+++ src/styles/global.css
@@ -1,0 +1,138 @@
+.slide-fade-enter-active {
+    transition: all .175s cubic-bezier(0.165, 0.84, 0.44, 1);
+}
+.slide-fade-leave-active {
+    transition: all .075s cubic-bezier(0.165, 0.84, 0.44, 1);
+}
+.slide-fade-enter, .slide-fade-leave-to
+    /* .slide-fade-leave-active for below version 2.1.8 */ {
+    transform: translateY(1.5vh);
+    opacity: 0;
+}
+
+.transition-all {
+    transition: all .225s cubic-bezier(0.165, 0.84, 0.44, 1);
+}
+
+.v-navigation-drawer {
+    transition: all .3s cubic-bezier(.25,.8,.5,1) !important;
+}
+
+.drawer-logo {
+    height: calc(256px + env(safe-area-inset-top));
+    padding: calc(32px + env(safe-area-inset-top)) 32px 32px 32px;
+    overflow: hidden;
+    transition: all .5s cubic-bezier(0.19, 1, 0.22, 1);
+}
+
+.drawer-logo:hover {
+    height: calc(304px + env(safe-area-inset-top));
+}
+.drawer-logo--two-line:hover {
+    height: calc(336px + env(safe-area-inset-top));
+}
+.drawer-logo > .description {
+    margin-top: 16px;
+    text-align: center;
+    line-height: 36px;
+    font-size: 24px;
+    opacity: 0;
+    transition: all .5s cubic-bezier(0.19, 1, 0.22, 1);
+}
+
+.drawer-logo:hover > .description {
+    opacity: 1;
+}
+
+.randomizedLogo > .v-image__image {
+    transition: background .5s cubic-bezier(0.19, 1, 0.22, 1);
+}
+
+.theme--light .bkop-light {
+    background: rgba(255, 255, 255, .75) !important;
+}
+
+.theme--dark .bkop-light {
+    background: rgba(46, 46, 46, .85) !important;
+}
+
+.theme--light .bkop-medium {
+    background: rgba(255, 255, 255, .9) !important;
+}
+
+.theme--dark .bkop-medium {
+    background: rgba(46, 46, 46, .9) !important;
+}
+
+.cursor-pointer {
+    cursor: pointer;
+}
+
+.transparentTable > .v-table__overflow > .v-table {
+    background: transparent;
+}
+
+.v-navigation-drawer::-webkit-scrollbar {
+    width: 2px;
+}
+
+.v-navigation-drawer::-webkit-scrollbar-thumb {
+    background-color: rgb(200, 200, 200);
+}
+
+/*.v-toolbar {*/
+/*  padding-top: env(safe-area-inset-top);*/
+/*}*/
+
+/*.v-footer {*/
+/*  height: calc(32px + env(safe-area-inset-bottom)) !important;*/
+/*  padding-bottom: calc(env(safe-area-inset-bottom));*/
+/*}*/
+
+.no-wrap--text {
+    word-break: break-word;
+}
+.no-break--text {
+    word-break: keep-all;
+}
+
+.monospace {
+    font-family: SF Mono, "Droid Sans Mono", Ubuntu Mono, Consolas, Courier New, Courier, monospace;
+}
+
+.footer--safe-area {
+    padding-top: 8px !important;
+    /*In case the old browsers doesn't support advanced CSS calculations*/
+    padding-bottom: 8px !important;
+    padding-bottom: calc(max(env(safe-area-inset-bottom), 8px)) !important;
+}
+
+/*.v-stepper__items, .v-stepper__wrapper {*/
+/*  overflow: initial !important;*/
+/*}*/
+
+.lang-ja .force-lang-font, .lang-ja {
+    /*font-family : 'ヒラギノ角ゴ ProN' , 'Hiragino Kaku Gothic ProN' , '游ゴシック' , '游ゴシック体' , YuGothic , 'Yu Gothic' , 'メイリオ' , Meiryo , 'ＭＳ ゴシック' , 'MS Gothic' , HiraKakuProN-W3 , 'TakaoExゴシック' , TakaoExGothic , 'MotoyaLCedar' , 'Droid Sans Japanese' , sans-serif !important;*/
+    font-family : 'ヒラギノ角ゴ ProN' , 'Hiragino Kaku Gothic ProN' , 'ヒラギノ明朝 ProN' , 'Hiragino Mincho ProN' , '游明朝','游明朝体',YuMincho,'Yu Mincho' , 'Meiryo', 'ＭＳ 明朝' , 'MS Mincho' , HiraMinProN-W3 , 'TakaoEx明朝' , TakaoExMincho , 'MotoyaLCedar' , 'Droid Sans Japanese' , serif !important;
+}
+.lang-ko .force-lang-font, .lang-ko {
+    font-family: 'Malgun Gothic', Gulim, 'Roboto', Tahoma, Arial, sans-serif !important
+}
+.force-not-lang-font {
+    font-family: Roboto, sans-serif!important
+}
+
+.x--safe-area {
+    padding-left: calc(max(env(safe-area-inset-left), 12px)) !important;
+    padding-right: calc(max(env(safe-area-inset-right), 12px)) !important;
+}
+
+.toolbar--safe-area {
+    padding-left: calc(max(env(safe-area-inset-left), 16px)) !important;
+    padding-right: calc(max(env(safe-area-inset-right), 16px)) !important;
+}
+
+.toolbar--safe-area .v-toolbar__content {
+    padding-left: 0 !important;
+    padding-right: 0 !important;
+}
