+++ src/views/About/Changelog.vue
@@ -73,7 +73,7 @@ export default {
           changes: [
             "添加：User ID 说明",
             "重构：更新日志页面",
-            "修复：有时无法自动隐藏 Crisp 的错误",
+            "修复：有时无法自动隐藏 CrispCustomizer 的错误",
             "修复：重新使用 SaaS 版 Sentry",
             "若干错误修复与性能优化"
           ]
